Added new inner Thread class to GLTextGridWindow to drive the VM in exported applications
The SwingWorker driver implementation would not execute if ran from an exported application; SwingWorkers are daemon threads and will stop if the main thread completes and no other non-daemon threads are running (ie: the Basic4GLj IDE window). To correct this the VM is run within a non-daemon thread when GLTextGridWindow is the main class.