added gradle task copyJarsForJPackage to prepare input directory for jpackage; refactored jpackage config to build app-image.
app-image may be overkill - originally configured to allow adding jars needed for debug and output processes at runtime, but app launchers cannot be added for those jars after the app-image is created which is why the new gradle task copyJarsForJPackage is needed. Keeping the app-image config around for now incase it's needed to embed samples and licenses. "--app-content" for JDK 21's jpackage may be usable in the future, but it doesn't appear to handle duplicates at the moment.